Changed home layout added discover fragment
Desc: now we have home fragment structure
