+++ .idea/vcs.xml
@@ -2,7 +2,5 @@
 <project version="4">
   <component name="VcsDirectoryMappings">
     <mapping directory="" vcs="Git" />
-    <mapping directory="$PROJECT_DIR$/AndroidAsync" vcs="Git" />
-    <mapping directory="$PROJECT_DIR$/ion" vcs="Git" />
   </component>
 </project>
