+++ app/src/main/res/drawable/circle_background.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android" >
+
+    <solid android:color="#FFFFFF" />
+    <corners
+        android:radius="999dp"
+        />
+
+</shape>
